requirements.txt
Pillow==5.2.0
python-magic==0.4.15
SQLAlchemy==1.2.11
